change `ScoringMethod` to abstract base class and rename it
