+++ src/nplinker/scoring/__init__.py
@@ -1,7 +1,7 @@
+from .abc import ScoringBase
 from .link_collection import LinkCollection
 from .metcalf_scoring import MetcalfScoring
-from .methods import ScoringMethod
 from .object_link import ObjectLink
 
 
-__all__ = ["LinkCollection", "MetcalfScoring", "ScoringMethod", "ObjectLink"]
+__all__ = ["LinkCollection", "MetcalfScoring", "ScoringBase", "ObjectLink"]
